Move Compute's spi package to compute.spi
