+++ gcloud-java-compute/src/main/java/com/google/gcloud/compute/Option.java
@@ -19,7 +19,7 @@ package com.google.gcloud.compute;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.base.MoreObjects;
-import com.google.gcloud.spi.ComputeRpc;
+import com.google.gcloud.compute.spi.ComputeRpc;
 
 import java.io.Serializable;
 import java.util.Objects;
